some left-over plot routine moved from objectives.py to plot.py; likely to be discarded in future
